Add Debug class and printTrustKeyStorePath method
Added a new class named Debug in the org.beanmaker.v2.util package. This class contains a method printTrustKeyStorePath which prints the trust/key store path from the 'javax.net.ssl.trustStore' system property, or the default path if the property is not set. It outputs these details to the provided PrintStream.