Add HttpConnections utility class in beanmaker2-util module
The new HttpConnections utility class contains methods for working with HTTP connections. These methods include functions for extracting cookies from an HTTP response and joining a map of cookies into a formatted string. The module-info.java file was also updated to include a dependency on java.net.http in the beanmaker2-util module.