Refactor methods and improve handling in email classes
The key changes include the correction of the method name `setHTmlContent` to `setHtmlContent` in `MessageBuilder` and `PlainMessageBuilder`, and the improvement in handling of `displayName` by using `Optional`. In addition, `PlainSender` now implements `Sender` and returns an instance of `SmtpServer` instead of `null` in `BasicSmtpServerBuilder`. The handling of addresses in `PlainRecipient` is also improved through method refactorization.