Refactor PlainMessageBuilder to support Image URLs and Files
The PlainMessageBuilder class in Beanmaker2-email, has been refurbished to support both Image URLs and Embedded Image Files, in which correspondingly the classes EmbeddedImageFile and EmbeddedImageUrl are created. Moreover, this refactor includes updates in the getString and validation methods to accomodate the two types of Embedded Images. Furthermore, these changes extend to interfacing the Message and MessageBuilder class, to define the newly added functionalities, as well as applying corresponding changes in other sections like BasicSmtpServerBuilder. Finally, the EmbeddedImage class is now EmbeddedImageUrl, reflecting the organization of image object types.