Fix a bug where empty non required label would not be ignored during bean creation/update
Optional labels are rare and were not treated properly. An empty reference need to be stored in the bean and not an empty label in the database. Code generation has been modified to archive this and also a new function has been added to DbBeanLabelEditor interface that will need to be implemented in client libraries/applications.