Add new method overloads with Boxed types and mark old versions Deprecated
Introduced method overloads that use Boolean, Integer, and Long types to replace those using primitive types. Marked the existing methods using primitives as @Deprecated for future removal. Ensured that null values are appropriately handled in the new methods.

This change was required to take into account optional values in beans that would create exception when master tables were displayed. Old versions of the function are kept for now to ensure binary compatibility with previously generated and compiled code. Ideally all projects using this version of the library should be recompiled.